Rework specification of SDK and associated dependencies
Ensure that the SDK toolchain is properly pushed through to rootfs.  Note
however that this commit needs the current version of rootfs (at least
commit 13a8096a).

The SDK is now either specified as a single SDK_ROOT key, as before, or as
individual keys for the separate BOOTGEN, BINUTILS_DIR, and possibly
SYSROOT components.